requirements.txt
# Flux LayerDiffuse Dependencies
# Based on: https://github.com/RedAIGC/Flux-version-LayerDiffuse

# Core diffusion libraries
diffusers==0.32.2
transformers>=4.48.0,<5.0.0

# Model handling and optimization
safetensors>=0.5.0
accelerate>=1.5.0
peft

# Image processing (compatible versions)
Pillow>=8.0.0,<12.0.0
opencv-python
numpy>=1.19.0,<3.0.0

# Utilities
tqdm
protobuf
sentencepiece

# Optional: For better performance
# torch>=2.0.0  # Usually provided by ComfyUI
# torchvision    # Usually provided by ComfyUI
# xformers       # For memory optimization (optional)

# Note: If you encounter dependency conflicts with other packages,
# you can try installing with --force-reinstall or create a separate environment
